requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.5
h11==0.16.0
idna==3.11
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
redis==7.0.1
requests==2.32.5
resend==2.27.0
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.49.3
stripe==15.0.1
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
vine==5.1.0
wcwidth==0.6.0
